# Optimize text layer: Avoid unneeded .textContent access